src/weather/requirements.txt
geopy
requests
meteostat==1.6.2
numpy
python-dateutil